add _Workflow.doughnut property that automatically generates the doughnut if is missing
