+++ nipoppy/workflows/__init__.py
@@ -1,4 +1,4 @@
 """Classes for running workflows on datasets."""
 
 from .dataset_init import InitWorkflow
-from .doughnut import DoughnutWorkflow, generate_doughnut, update_doughnut
+from .doughnut import DoughnutWorkflow
